#312 started to implement forgot password page. Also moved classes from main.dart to corresponding files.
