+++ ludos/mobile/lib/game_page_widget.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class GamePageWidget extends StatefulWidget {
+  const GamePageWidget({super.key});
+
+  @override
+  State<GamePageWidget> createState() => _GamePageWidgetState();
+}
+
+class _GamePageWidgetState extends State<GamePageWidget> {
+  @override
+  Widget build(BuildContext context) {
+    return const Placeholder();
+  }
+}
+
+
